requirements.txt
huggingface-hub==0.26.0
hydra-core==1.3.2
librosa==0.9.2
matplotlib==3.9.2
matplotlib-inline==0.1.7
numpy==1.23.5
pandas==2.2.3
pytorch-metric-learning==2.6.0
scikit-learn==1.5.2
scipy==1.14.1
sentencepiece==0.1.97
soundfile==0.12.1
timm==1.0.11
tokenizers==0.20.1
torch==2.4.1
torch-complex==0.4.4
torchaudio==2.4.1
torchvision==0.19.1
tqdm==4.66.5
transformers==4.45.2
jsonlines==4.0.0